Refactor state.rs away and rename the trait to GameState.
